Extract MimeTypeSelect to component
